release (major): improvements and refactoring
Split command flags into what should have been multiple command modes.
This allowed for a refactor of the code into cleaner branches of
operation.
Made a number of fixes on how paths were handled that broke the script
on windows. It is now also usable from Windows.
Under inspect, default was N but user was told default was Y. Fixed
that.